[INSTX-7939] Fixes for MSVC

+++ tests/BarcodeClassifierTest.cpp
@@ -301,7 +301,8 @@ CATCH_TEST_CASE(
 
     read->read_common.num_trimmed_samples = 0;
 
-    auto records = read->read_common.extract_sam_lines(true /* emit moves*/, 10, false);
+    auto records = read->read_common.extract_sam_lines(true /* emit moves*/,
+                                                       static_cast<uint8_t>(10), false);
 
     // Push a Read type.
     pipeline->push_message(std::move(read));
@@ -379,7 +380,7 @@ CATCH_TEST_CASE(
             // Number of trimmed bases is 100, so number of moves should be 2 * 100.
             CATCH_CHECK(read_common.get_raw_data_samples() == 100 * 2 * stride);
 
-            auto bams = read_common.extract_sam_lines(0, 10, false);
+            auto bams = read_common.extract_sam_lines(0, static_cast<uint8_t>(10), false);
             auto& rec = bams[0];
             auto [mod_str, mod_probs] = dorado::utils::extract_modbase_info(rec.get());
         }
