+++ tests/ReadTest.cpp
@@ -244,7 +244,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
         // Test generation
         const char* expected_methylation_tag_10_score = "A+a.,0,1;C+m.,1,0;";
         std::vector<int64_t> expected_methylation_tag_10_score_prob{20, 254, 252, 252};
-        auto lines = read_common.extract_sam_lines(false, 10, false);
+        auto lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(10), false);
         CATCH_REQUIRE(!lines.empty());
         bam1_t* aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")),
@@ -255,7 +255,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
         // Test generation at higher rate excludes the correct mods.
         const char* expected_methylation_tag_50_score = "A+a.,2;C+m.,1,0;";
         std::vector<int64_t> expected_methylation_tag_50_score_prob{254, 252, 252};
-        lines = read_common.extract_sam_lines(false, 50, false);
+        lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(50), false);
         CATCH_REQUIRE(!lines.empty());
         aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")),
@@ -266,7 +266,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
         // Test generation at max threshold rate excludes everything
         const char* expected_methylation_tag_255_score = "A+a.;C+m.;";
         std::vector<int64_t> expected_methylation_tag_255_score_prob{};
-        lines = read_common.extract_sam_lines(false, 255, false);
+        lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(255), false);
         CATCH_REQUIRE(!lines.empty());
         aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")),
@@ -284,7 +284,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
 
         read_common.mod_base_info = std::make_shared<dorado::ModBaseInfo>(modbase_alphabet_CHEBI,
                                                                           modbase_long_names, "");
-        auto lines = read_common.extract_sam_lines(false, 50, false);
+        auto lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(50), false);
         CATCH_REQUIRE(!lines.empty());
         bam1_t* aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")), Equals(expected_methylation_tag_CHEBI));
@@ -299,7 +299,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
         read_common.mod_base_info = std::make_shared<dorado::ModBaseInfo>(
                 modbase_alphabet, modbase_long_names, context);
 
-        auto lines = read_common.extract_sam_lines(false, 10, false);
+        auto lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(10), false);
         CATCH_REQUIRE(!lines.empty());
         bam1_t* aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")),
@@ -316,7 +316,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
         read_common.mod_base_info = std::make_shared<dorado::ModBaseInfo>(
                 modbase_alphabet, modbase_long_names, context);
 
-        auto lines = read_common.extract_sam_lines(false, 10, false);
+        auto lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(10), false);
         CATCH_REQUIRE(!lines.empty());
         bam1_t* aln = lines[0].get();
         CATCH_CHECK_THAT(bam_aux2Z(bam_aux_get(aln, "MM")),
@@ -332,7 +332,7 @@ CATCH_TEST_CASE(TEST_GROUP ": Methylation tag generation", TEST_GROUP) {
 
         read_common.mod_base_info = std::make_shared<dorado::ModBaseInfo>(modbase_alphabet_unknown,
                                                                           modbase_long_names, "");
-        auto lines = read_common.extract_sam_lines(false, 50, false);
+        auto lines = read_common.extract_sam_lines(false, static_cast<uint8_t>(50), false);
         CATCH_REQUIRE(!lines.empty());
         bam1_t* aln = lines[0].get();
         CATCH_CHECK(bam_aux_get(aln, "MM") == NULL);
